update: main.js not to kill pid -1; update: new package-lock.json

+++ nft_generator_electron/main.js
@@ -47,8 +47,10 @@ app.whenReady().then(() => {
 });
 
 app.on("before-quit", (event) => {
-    process.kill(pid, "SIGINT");
-    console.log("NFTG Server Killed");
+    if (pid != -1) {
+        process.kill(pid, "SIGINT");
+        console.log("NFTG Server Killed");
+    }
 });
 
 ipcMain.handle("pid", (event, _pid) => {
